Changed ints to longs. Fixed IngredientService Tests.

+++ src/main/java/com/charlie/morerecipes/domain/Recipe.java
@@ -1,7 +1,5 @@
 package com.charlie.morerecipes.domain;
 
-import org.hibernate.engine.internal.Cascade;
-
 import javax.persistence.*;
 import java.util.HashSet;
 import java.util.Set;
@@ -11,7 +9,7 @@ public class Recipe {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String description;
     private Integer prepTime;
@@ -43,11 +41,11 @@ public class Recipe {
     public Recipe() {
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
