+++ src/test/java/com/charlie/morerecipes/domain/CategoryTest.java
@@ -16,7 +16,7 @@ public class CategoryTest {
 
     @Test
     public void getId() {
-        Integer idValue = 4;
+        Long idValue = 4L;
         category.setId(idValue);
         assertEquals(idValue,category.getId());
     }
